% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_columns.R
\name{all_columns}
\alias{all_columns}
\title{Retrieves all column names from different csv files}
\usage{
all_columns(csv_files, ...)
}
\arguments{
\item{csv_files}{a character vector with paths to one ore more csv files}

\item{...}{further arguments to be passed to data.table::fread() (except nrows, which is set to 0)}
}
\description{
Retrieves all column names from different csv files
}
\examples{

# The package contains two small csv files: csv01.csv and csv02.csv
# The first one has 3 columns: x, y and z
# The second one has only 2 columns: x and z
all_columns(
  c(system.file("extdata", "csv01.csv", package = "impexpcsv"),
  system.file("extdata", "csv02.csv", package = "impexpcsv"))  
)

}
