% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_csv.R
\name{bind_csv}
\alias{bind_csv}
\title{Create a "row-binded" csv file appending common columns of various csv files}
\usage{
bind_csv(csv_files, col_names = NULL, csv_output, sep_output = ";", ...)
}
\arguments{
\item{csv_files}{a character vector with paths to one ore more csv files}

\item{col_names}{a character vector with the names of the columns to be imported from each csv file. If NULL (default), all common columns will be imported.}

\item{csv_output}{path to the csv file to be created and saved}

\item{sep_output}{column separator of csv_output (defalut: ";")}

\item{...}{further arguments to be passed to data.table::fread() (except nrows, which is set to 0)}
}
\value{
a NULL value. A summary message is printed on the screen.
}
\description{
Create a "row-binded" csv file appending common columns of various csv files
}
\examples{
\dontrun{
csv_files <- c(
  system.file("extdata", "csv01.csv", package = "impexpcsv"),
  system.file("extdata", "csv02.csv", package = "impexpcsv")  
)

bind_csv(csv_files, csv_output = "big_csv.csv")
bind_csv(csv_files, col_names = "x", csv_output = "big_csv.csv")
bind_csv(csv_files, col_names = c("z", "x"), csv_output = "big_csv.csv")
}

}
