% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_columns.R
\name{common_columns}
\alias{common_columns}
\title{Find common columns of various csv files}
\usage{
common_columns(csv_files, ...)
}
\arguments{
\item{csv_files}{a character vector with paths to one ore more csv files}

\item{...}{further arguments to be passed to data.table::fread() (except nrows, which is set to 1)}
}
\value{
a character vector with the names of the common columns of the csv files
}
\description{
Find common columns of various csv files
}
\examples{

# The package contains two small csv files: csv01.csv and csv02.csv
# The first one has 3 columns: x, y and z
# The second one has only 2 columns: x and z
common_columns(
  c(system.file("extdata", "csv01.csv", package = "impexpcsv"),
  system.file("extdata", "csv02.csv", package = "impexpcsv"))  
)
}
