% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impexpcsv-package.R
\docType{package}
\name{impexpcsv-package}
\alias{impexpcsv}
\alias{impexpcsv-package}
\title{impexpcsv: Import Various csv Files with Common Columns and Export in One csv file}
\description{
Import various csv files with common columns and export in one csv file. Csv files are imported with data.table::fread(), row-binded with data.table::rbindlist() and exported with data.table::fwrite().
}
\keyword{internal}
