% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_impexpcsv.R
\docType{package}
\name{impexpcsv}
\alias{impexpcsv}
\title{Import Various csv Files with Common Columns and Export in One csv file}
\description{
Import various csv files with common columns and export in one
csv file. Csv files are imported with data.table::fread(),
row-binded with data.table::rbindlist() and exported with
data.table::fwrite().
}
