% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_list_csv.R
\name{import_list_csv}
\alias{import_list_csv}
\title{Import various csv files in a "row-binded" data.table with the same columns}
\usage{
import_list_csv(csv_files, col_names = NULL, ...)
}
\arguments{
\item{csv_files}{a character vector with paths to one ore more csv files}

\item{col_names}{a character vector with the names of the columns to be imported from each csv file. If NULL (default), all common columns will be imported.}

\item{...}{further arguments to be passed to data.table::fread() (except nrows, which is set to 0)}
}
\value{
a data table with all the csv files binded together
}
\description{
Import various csv files in a "row-binded" data.table with the same columns
}
\examples{
# The package contains two small csv files: csv01.csv and csv02.csv
# The first one has 3 columns: x, y and z
# The second one has only 2 columns: x and z

# By default, import_list_csv imports all the common columns of the csv 
# files
import_list_csv(
  c(system.file("extdata", "csv01.csv", package = "impexpcsv"),
  system.file("extdata", "csv02.csv", package = "impexpcsv"))  
)

# Import only column "x" of each csv file
import_list_csv(
  c(system.file("extdata", "csv01.csv", package = "impexpcsv"),
  system.file("extdata", "csv02.csv", package = "impexpcsv")),
  col_names = "x"
)

}
